+++ src/Resources/Resources.Test/ScenarioTests/ResourceGroupTests.ps1
@@ -316,6 +316,11 @@ function Test-FindResourceGroup
 		# Assert
 		Assert-AreEqual @($expected3).Count $expectedCount
 
+		# Assert
+		Assert-ThrowsContains `
+            { Get-AzResourceGroup -Name "https://RG1?default" } `
+            "Resource groups with wildcard pattern 'https://RG1?default' do not exist."
+
 		$expected4 = Get-AzResourceGroup -Tag @{ testtag = $null}
         # Assert
         Assert-AreEqual @($expected4).Count 2